pyproject.toml
[project]
name = "open-secrets-press"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "dagster>=1.11.13",
    "dagster-webserver>=1.11.13",
    "duckdb>=1.4.0",
    "en-core-web-sm",
    "lxml>=6.0.2",
    "lxml-html-clean>=0.4.2",
    "mediacloud>=4.4.0",
    "newspaper3k>=0.2.8",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "spacy>=3.8.7",
]

[tool.dagster]
module_name = "data.orchestration.definitions"

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
